requirements.txt
# FastAPI
fastapi==0.65.2

# Prometheus
prometheus-client==0.11.0

# Uvicorn
uvicorn==0.13.4

# PyYAML
pyyaml==5.4.1